Move tests out of src folder
