+++ jest.config.js
@@ -1,7 +1,7 @@
 module.exports = {
     preset: 'ts-jest',
     testEnvironment: 'node',
-    testMatch: ['/**/src/test/**/*.test.ts'],
+    testMatch: ['/**/test/**/*.test.ts'],
     collectCoverage: true,
     collectCoverageFrom: ['src/**/*.ts', 'run'],
     coverageReporters: ['json'],
