+++ src/commands/test.ts
@@ -3,6 +3,11 @@ import { Logger } from '../logUtils';
 import { createCoverageMap } from 'istanbul-lib-coverage';
 
 export async function runTests() {
+
+  if (!process.env.GITHUB_TOKEN) {
+    throw new Error('Please set the GITHUB_TOKEN environmental variable');
+  }
+
   try {
     exec('npx jest --json --coverage --silent', (err, stdout, stderr) => {
         if (err) {
@@ -31,8 +36,9 @@ export async function runTests() {
         console.log(`Coverage: ${Math.round(averageCoverage)}%`);
         console.log(`${passed}/${total} test cases passed. ${Math.round(averageCoverage)}% line coverage achieved.`);
     });
-  } catch (error) {
+  } catch (error: any) {
     Logger.logInfo("Error while running runTests()")
-    Logger.logDebug(`${error}`);
+    Logger.logDebug(error);
+    throw new Error("Error while running jest. Check logs.")
   }
 }
