+++ test/install.test.ts
